--- api/src/opentrons/drivers/asyncio/magdeck/__init__.py
@@ -1,9 +1,0 @@
-from .driver import MagDeckDriver
-from .simulator import SimulatingDriver
-from .abstract import AbstractMagDeckDriver
-
-__all__ = [
-    "AbstractMagDeckDriver",
-    "SimulatingDriver",
-    "MagDeckDriver"
-]
