+++ api/src/opentrons/drivers/mag_deck/__init__.py
@@ -1,7 +1,10 @@
-from opentrons.drivers.mag_deck.driver import MagDeck, SimulatingDriver
+from .abstract import AbstractMagDeckDriver
+from .simulator import SimulatingDriver
+from .driver import MagDeckDriver
 
 
 __all__ = [
-    'MagDeck',
-    'SimulatingDriver'
+    "AbstractMagDeckDriver",
+    "SimulatingDriver",
+    "MagDeckDriver"
 ]
