+++ api/src/opentrons/hardware_control/emulation/magdeck.py
@@ -5,7 +5,7 @@ The purpose is to provide a fake backend that responds to GCODE commands.
 
 import logging
 from typing import Optional
-from opentrons.drivers.asyncio.magdeck.driver import GCODE
+from opentrons.drivers.mag_deck.driver import GCODE
 from opentrons.hardware_control.emulation.parser import Parser, Command
 from .abstract_emulator import AbstractEmulator
